eval/mcp-eval/requirements.txt
mcp-eval>=0.0.1
pyyaml>=6.0
